requirements.txt
flask==2.3.3
PyPDF2==3.0.1
google-generativeai==0.3.1
langchain==0.1.1
langchain-google-genai==0.0.5
faiss-cpu==1.7.4
python-dotenv==0.19.1
